removed profile page
