+++ .env.example
@@ -15,3 +15,5 @@ CLOUDINARY_URL=
 CLOUDINARY_CLOUD_NAME=""  
 CLOUDINARY_API_KEY=""  
 CLOUDINARY_API_SECRET=""  
+NEXT_PUBLIC_BICONOMY_DASH_API=
+NEXT_PUBLIC_BICONOMY_API=
