+++ .gitignore
@@ -7,7 +7,7 @@
 v2mp.code-workspace
 # testing
 /coverage
-
+archives/
 # next.js
 /.next/
 /out/
